app/src/main/java/com/dicoding/storyapp/ui/maps/MapsViewModel.kt
package com.dicoding.storyapp.ui.maps

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import com.dicoding.storyapp.data.Result
import com.dicoding.storyapp.data.StoryRepository
import com.dicoding.storyapp.data.source.remote.response.GetAllStoriesResponse
import dagger.hilt.android.lifecycle.HiltViewModel
import javax.inject.Inject

@HiltViewModel
class MapsViewModel @Inject constructor(private val storyRepository: StoryRepository) :
    ViewModel() {

    fun getAllStoriesWithLocation(token: String?): LiveData<Result<GetAllStoriesResponse>> {
        return storyRepository.getAllStoriesWithLocation(token)
    }
}